powerpc: Set per-task key for user ROP protection
The DEXCR Non-privileged Hash Instruction Enable controls whether the
POWER ISA v3.1 'hashst' and 'hashchk' instructions are NOPs or not.
These instructions implement userspace ROP protection by storing a hash
of the link register and a secret, per-thread HASHKEY on function entry
and recomputing and checking the hash on function exit. Implement the
initialization and context restore of the per-thread HASHKEY.

Signed-off-by: Christopher M. Riedl <cmr@bluescreens.de>