Documentation: Document PowerPC kernel DEXCR interface
Describe the DEXCR and document how to configure it.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

---

v3:	* Revise (H)DEXCR width to 64 bits
	* Revise configuration section
v2:	* Document coredump & ptrace support
v1:	* Remove the dynamic control docs, describe the static config
	  option

This documentation is a little bare for now, but will be expanded on
when dynamic DEXCR control is added.