selftests/powerpc: Add generic read/write file util
File read/write is reimplemented in about 5 different ways in the
various PowerPC selftests. This indicates it should be a common util.

Add a common read_file / write_file implementation and convert users
to it where (easily) possible.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>

---

v4:	* Set errno as expected by many users that call perror() if
	  it fails (e.g., read_debugfs_file users in powerpc/security)
	* misc refactoring