powerpc/dexcr: Initialise the DEXCR based on firmware
Currently the DEXCR is set to a fixed value on boot. To help future
proof kernels for new aspects, split the init value into explicit
set and clear bitmasks. We currently set NPHIE and do not clear
anything.

Also include the required synchronization sequence for changing the
kernel DEXCR value.

Signed-off-by: Benjamin Gray <bgray@linux.ibm.com>